# backend/internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "STA Platform API",
    "version": "v1",
    "description": "Auto-generated shallow spec for the STA Go backend. Method, path, tags and security are accurate; request/response schemas are generic except for a few hand-augmented endpoints. Regenerate with `python3 docs/openapi/generate.py`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "local"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sta_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "adminMFA": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MFA-Code"
      },
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "extraction / agent service token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/api/v1/admin/ingestion/brochure-candidates/{candidateID}/review": {
      "post": {
        "operationId": "post_api_v1_admin_ingestion_brochure_candidates_candidateID_review",
        "summary": "Create admin ingestion brochure candidates candidateID review",
        "tags": [
          "admin-ingestion"
        ],
        "parameters": [
          {
            "name": "candidateID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ingestion/brochure-runs": {
      "get": {
        "operationId": "get_api_v1_admin_ingestion_brochure_runs",
        "summary": "Get admin ingestion brochure runs",
        "tags": [
          "admin-ingestion"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ingestion/brochure-runs/{runID}": {
      "get": {
        "operationId": "get_api_v1_admin_ingestion_brochure_runs_runID",
        "summary": "Get admin ingestion brochure runs runID",
        "tags": [
          "admin-ingestion"
        ],
        "parameters": [
          {
            "name": "runID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ingestion/brochure-runs/{runID}/review": {
      "post": {
        "operationId": "post_api_v1_admin_ingestion_brochure_runs_runID_review",
        "summary": "Create admin ingestion brochure runs runID review",
        "tags": [
          "admin-ingestion"
        ],
        "parameters": [
          {
            "name": "runID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ingestion/jobs/{jobID}": {
      "get": {
        "operationId": "get_api_v1_admin_ingestion_jobs_jobID",
        "summary": "Get admin ingestion jobs jobID",
        "tags": [
          "admin-ingestion"
        ],
        "parameters": [
          {
            "name": "jobID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ingestion/jobs/{jobID}/retry": {
      "post": {
        "operationId": "post_api_v1_admin_ingestion_jobs_jobID_retry",
        "summary": "Create admin ingestion jobs jobID retry",
        "tags": [
          "admin-ingestion"
        ],
        "parameters": [
          {
            "name": "jobID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}